requirements.txt
torch>=2.2
transformers>=4.42
peft>=0.11
accelerate>=0.30
datasets>=2.20
sentencepiece>=0.1.99
numpy>=1.26
scikit-learn>=1.4
tqdm>=4.66
pyyaml>=6.0
